requirements.txt
ccxt
python-dotenv
Flask
gunicorn 